feat: 메인 페이지 스켈레톤 적용
