+++ frontend2/package.json
@@ -13,6 +13,7 @@
     "react-cookie": "^4.1.1",
     "react-dom": "^18.2.0",
     "react-loader-spinner": "^5.3.4",
+    "react-loading-skeleton": "^3.3.1",
     "react-redux": "^8.1.1",
     "react-router-dom": "^6.14.1",
     "react-scripts": "5.0.1",
