+++ frontend2/src/components/molecules/InputGroup.jsx
@@ -1,11 +1,11 @@
 import React from "react";
 import Box from "../atoms/Box";
 
-export default function InputGroup({ label, value, disabled }) {
+export default function InputGroup({ children, value, disabled }) {
   return (
     <Box>
       <label>
-        {label}
+        {children}
         <input type="text" value={value} disabled={disabled} />
       </label>
     </Box>
